requirements.txt
python-dotenv~=0.17.1
discord.py>=1.7.1